seperated redis cache methods for it to work
