+++ Projects/kz-srikar/PaginationBackend/src/main/resources/application.properties
@@ -7,4 +7,4 @@ spring.datasource.password=123235Mv@
 
 spring.cache.type=redis
 spring.redis.host=localhost
-spring.redis.port=6379
+spring.redis.port=6360
